Update runG3Download.sh and variant_effect_predictor
runG3Download.sh refactored, corrected coding bugs in singleDownloadGET()
function.
Expanded variant_effect_predictor optional parameters.

+++ DNA/GDC_dna_seq/widgets/GDC_dna_seq/gen3_download/Dockerfiles/runG3Download.sh
@@ -3,6 +3,7 @@
 #things to do - can obtain manifest from endpoint for any guid to get file and md5 hash so we can verify before proceeding
 
 gdcapiurl='https://api.gdc.cancer.gov'
+exitCode=0
 
 #create a temporary directory for API download and mv the downloads when complete an no error
 #remove the temporary directory if there is a interrupt or error
@@ -10,12 +11,10 @@ gdcapiurl='https://api.gdc.cancer.gov'
 #need to also check for gen3 errors
 
 function cleanup(){
- #use global tempDir as this will depend on the GUID being downloaded or the manifest being downloaded
- #make sure that basename tempdir starts with temp so that we don't remove something disastrous with a typo
- basetemp=$(basename tempDir) 
- if [ ${basetemp:0:4} == 'temp' ]; then
-	rm -rf $tempDir
- fi
+	#use global tempDir as this will depend on the GUID being downloaded or the manifest being downloaded
+	#make sure that basename tempdir starts with temp so that we don't remove something disastrous with a typo
+	basetemp=$(basename $tempDir)
+	[[ -n "$basetemp" && "${basetemp:0:4}" = "temp" ]] && rm -rf $tempDir
 }
 
 #call cleanup if we exit not so nicely
@@ -23,157 +22,135 @@ trap "cleanup  -1 " SIGINT INT TERM
 
 #there is a bug with the multi-download api in that the files are not properly tarballed - so use only singleDownload endpoint
 function singleDownloadGET(){
- local myid=$1
- downloaded=0
- cd $tempDir; curl -OJ --header "X-Auth-Token: $token" $gdcapiurl/data/$myid
- if [[ $? != 0 ]]; then
-   echo "curl error $?"
-   cleanup
-   return
- else
-   if [ -f "$tempDir/$guid" ]; then
-     cat $tempDir/$guid
-     cleanup
-     return
-   fi
-   if [ -f "$tempDir/$data" ]; then
-     cat $tempDir/data
-     cleanup
-     return
-   fi  
-   if [ -z "$(ls -A $tempDir)" ]; then
-     cleanup
-     return
-   fi
-   mv $tempDir/* $outputDir/
-   downloaded=1
-   cleanup
-   return     
- fi
+	local myid=$1
+	local downloaded=0
+	tempDir=$(mktemp -d -p $downloadDir -t tempXXXXXX)
+	[ -z "$tempDir" ] && return 1
+	pushd $tempDir > /dev/null
+	if ! curl -OJ --header "X-Auth-Token: $token" $gdcapiurl/data/$myid; then
+		echo "curl error $?"
+	elif [ -f "$tempDir/$myid" ]; then
+		cat $tempDir/$myid
+	elif [ -f "$tempDir/data" ]; then
+		cat $tempDir/data
+	elif [ -n "$(ls -A $tempDir/*tar)" ]; then
+		# download successful
+		if [[ -n "$untarfiles" && "$untarfiles" = "True" ]]; then
+			tar -xf $tempDir/* -C $downloadDir
+		else
+			mv $tempDir/* $downloadDir
+		fi
+		downloaded=1
+	fi
+	popd > /dev/null
+	cleanup
+	return $downloaded
 }
+
 function convertManifest(){
- echo "tail -n +2 $manifest | awk '{printf "%s ",$1}'"
- guidsArray=($(tail -n +2 $manifest | awk '{printf "%s ",$1}'))
+	echo "tail -n +2 $manifest | awk '{printf "%s ",$1}'"
+	guidsArray=($(tail -n +2 $manifest | awk '{printf "%s ",$1}'))
 }
 
 function downloadWithToken(){
-  if [ -n "$manifest" ]; then
-    convertManifest
-  else
-    guidsArray=($(convertJsonToArrayNoQuotes $guids))
-  fi
-  myArray=()
-  for guid in "${guidsArray[@]}"; do
-    tempDir=$outputDir/temp$guid
-    mkdir -p $tempDir
-    singleDownloadGET $guid
-    if [ $downloaded == 0 ];then
-      myArray+=($guid)  
-    fi
-  done
-  guidsArray=(${myArray[@]})
+	if [ -n "$manifest" ]; then
+		convertManifest
+	else
+		guidsArray=($(convertJsonToArrayNoQuotes $guids))
+	fi
+	local myArray=()
+	for guid in "${guidsArray[@]}"; do
+		singleDownloadGET $guid && myArray+=($guid)
+	done
+	guidsArray=(${myArray[@]})
+	return ${#myArray[@]}
 }
 
 function convertJsonToArrayNoQuotes(){
- #echos out a string that can be converted to a bash array to get around not being able to return string or array	
- local string=$1
- if [ ${string:0:1} = '[' ];then
-  #remove square brackets at beginning and end with substring
-  #put spaces with , replacement
-  #remove all " in this case
-  string=$(echo "${string:1:${#string}-2}" | sed 's/\,/ /g' | sed 's/\"/ /g')
- fi
- echo "$string"
+	#echos out a string that can be converted to a bash array to get around not being able to return string or array
+	local string=$1
+	if [ ${string:0:1} = '[' ];then
+		#remove square brackets at beginning and end with substring
+		#put spaces with , replacement
+		#remove all " in this case
+		string=$(echo "${string:1:${#string}-2}" | sed 's/\,/ /g' | sed 's/\"/ /g')
+	fi
+	echo "$string"
 }
 
 function singleDownload(){
- exitCode=0
- if [ -z $guidsArray ]; then
-   guidsArray=($(convertJsonToArrayNoQuotes $guids))
- fi
- for guid in "${guidsArray[@]}"; do
-   echo "gen3-client download-single --profile=$profile --no-prompt --guid=$guid ${flags[@]}"
-   gen3-client download-single --profile=$profile --no-prompt --guid=$guid ${flags[@]} 2> >(tee -a /tmp/log$guid >&2)
-   errors=$(fgrep 'Details of error:' /tmp/log$guid)
-   if [ -n "$errors" ]; then
-     echo "Exiting with error $errors"
-     exitCode=1
-     return 
-   fi
- done
+	[ -z $guidsArray ] && guidsArray=($(convertJsonToArrayNoQuotes $guids))
+	for guid in "${guidsArray[@]}"; do
+		echo "gen3-client download-single --profile=$profile --no-prompt --guid=$guid ${flags[@]}"
+		gen3-client download-single --profile=$profile --no-prompt --guid=$guid ${flags[@]} 2> >(tee -a /tmp/log$guid >&2)
+		errors=$(fgrep 'Details of error:' /tmp/log$guid)
+		if [ -n "$errors" ]; then
+			echo "Exiting with error $errors"
+			exitCode=1
+			return
+		fi
+	done
 }
+
 function multiDownload(){
-  exitCode=0
-  echo "Downloading using manifest"	
-  echo "gen3-client download-multiple --profile=$profile --no-prompt ${flags[@]}"
-  gen3-client download-multiple --profile=$profile --no-prompt ${flags[@]} 2> >(tee -a /tmp/logManifest >&2)
-  errors=$(fgrep 'Details of error:' /tmp/logManifest)
-  if [ -n "$errors" ]; then
-    echo "Exiting with error $errors"
-    exit 1
-  fi 
+	echo "Downloading using manifest"
+	echo "gen3-client download-multiple --profile=$profile --no-prompt ${flags[@]}"
+	gen3-client download-multiple --profile=$profile --no-prompt ${flags[@]} 2> >(tee -a /tmp/logManifest >&2)
+	errors=$(fgrep 'Details of error:' /tmp/logManifest)
+	if [ -n "$errors" ]; then
+		echo "Exiting with error $errors"
+		exitCode=1
+	fi
 }
 
 #check if both guid and manifest given
 #For now force user to use one or the other - otherwise this may cause difficulties with auto-multithread with the manifest being downloaded each time
 #they can use two instances of the widget if they really want to do this and not merge the manifest
 #if we pass a flag to indicate multi-thread execution (which we may) then we might modify this
-
 if [[ -n $guids && -n $manifest ]]; then
-  echo "Please choose either a GUID or a manifest file not both"
-  echo "You can merge the GUID into the manifest or use two instances of the widget to download both"
-  exit 1
+	echo "Please choose either a GUID or a manifest file not both"
+	echo "You can merge the GUID into the manifest or use two instances of the widget to download both"
+	exit 1
 fi
-
 
 #First try with the old api
-if [ -z $gdctoken ]; then
-   echo "no gdc token given - use gen3 fence to download"
-else
-  if [ -f $gdctoken ]; then
-    token=$(cat $gdctoken)
-    downloadWithToken
-  fi
-  #if there are no files left in guidsArray then all files have been successfully downloaded - exit
-  (( ${#guidsArray[@]} )) || exit $exitCode
+if [[ -n "$gdctoken" && -f "$gdctoken" ]]; then
+	token=$(cat $gdctoken)
+	#if there are no files left in guidsArray then all files have been successfully downloaded - exit
+	downloadWithToken && exit 0
 fi
 
+echo "no gdc token given - use gen3 fence to download"
 echo "Attempting to authenticate using gen3 fence service"
 #Authenticate the container by copying or creating a config
 #the commons might change so we do not hardcode the url
-if [ -z $datacommons_url ]; then
-	datacommons_url="https://nci-crdc.datacommons.io/"
-fi
+[ -z $datacommons_url ] && datacommons_url="https://nci-crdc.datacommons.io/"
 
 #assume config file basename is config when given by user
 #otherwise assume that the file given by user is a credentials file
 
 credBasename=$(basename $cred)
 if [ $credBasename == "config" ]; then
-    mkdir -p /root/.gen3
-   	cp credBasename /root/.gen3/config
-else
-  gen3-client configure --profile=$profile --cred=$cred --apiendpoint=$datacommons_url
-  if [ $? != 0 ]; then
-    echo "was not successful in creating new profile"
-    exit 1
-  fi
+	mkdir -p /root/.gen3
+	cp $cred /root/.gen3/config
+elif ! gen3-client configure --profile=$profile --cred=$cred --apiendpoint=$datacommons_url; then
+	echo "was not successful in creating new profile"
+	exit 1
 fi
 
 #check if we have a configuration defined now
 if [ -f "/root/.gen3/config" ]; then
- #now we can begin download
- flags=( "$@" ) 
- gen3-client auth --profile=$profile
- if [ -z $manifest ]; then 
-	singleDownload
- else
-    multiDownload
- fi
+	#now we can begin download
+	flags=( "$@" )
+	gen3-client auth --profile=$profile
+	if [ -z $manifest ]; then
+		singleDownload
+	else
+		multiDownload
+	fi
 else
-  echo "must provide a valid config or credentials file" 
-  exit 1 
+	echo "must provide a valid config or credentials file"
+	exit 1
 fi
+
 exit $exitCode
-
-
